Refactor compiler structure
Previously the AppsCompiler class had too many responsibilities and
concerns, so I've split it into a few more, more simply scoped classes.

This is ground leveling to introduce rollup
