+++ src/misc/appPackager.ts
@@ -5,8 +5,7 @@ import * as Yazl from 'yazl';
 import glob, { IOptions } from 'glob';
 
 import { FolderDetails } from './folderDetails';
-import { AppsCompiler } from '../AppsCompiler';
-import { ICompilerDescriptor } from '../definition';
+import { ICompilerDescriptor, ICompilerResult } from '../definition';
 
 export class AppPackager {
     public static GlobOptions: IOptions = {
@@ -27,7 +26,12 @@ export class AppPackager {
 
     private zip: Yazl.ZipFile = new Yazl.ZipFile();
 
-    constructor(private readonly compilerDesc: ICompilerDescriptor, private fd: FolderDetails, private compiledApp: AppsCompiler, private outputFilename: string) { }
+    constructor(
+        private readonly compilerDesc: ICompilerDescriptor,
+        private fd: FolderDetails,
+        private compilationResult: ICompilerResult,
+        private outputFilename: string,
+    ) { }
 
     public async zipItUp(): Promise<string> {
         const zipName = this.outputFilename;
@@ -36,22 +40,21 @@ export class AppPackager {
 
         this.overwriteAppManifest();
 
-        this.zipFromCompiledSource();
+        this.zipFilesFromCompiledSource();
 
         await this.zipSupportFiles();
 
         this.zip.end();
 
-        await this.asyncWriteZip(this.zip, zipName);
+        await this.writeZip(this.zip, zipName);
 
         return zipName;
     }
 
     private overwriteAppManifest(): void {
-        const { AppInterface } = this.compiledApp.appRequire('@rocket.chat/apps-engine/definition/metadata');
-
-        this.fd.info.implements = this.compiledApp.getImplemented()
-            .filter((interfaceName) => !!AppInterface[interfaceName]) as Array<typeof AppInterface>;
+        // At this point, the interface names in the implemented property
+        // have been validated and guaranteed to be correct, so we type cast
+        this.fd.info.implements = this.compilationResult.implemented as any;
 
         fs.writeFileSync(this.fd.infoFile, JSON.stringify(this.fd.info, null, 4));
     }
@@ -87,9 +90,13 @@ export class AppPackager {
             }));
     }
 
-    private zipFromCompiledSource(): void {
-        Object.entries(this.compiledApp.output())
-            .map(([filename, contents]) => this.zip.addBuffer(Buffer.from(contents), filename, { compress: true }));
+    private zipFilesFromCompiledSource(): void {
+        Object.entries(this.compilationResult.files)
+            .map(([filename, contents]) => this.zip.addBuffer(
+                Buffer.from(contents.compiled),
+                filename,
+                { compress: true },
+            ));
     }
 
     // tslint:disable-next-line:promise-function-async
@@ -107,10 +114,12 @@ export class AppPackager {
     }
 
     // tslint:disable-next-line:promise-function-async
-    private asyncWriteZip(zip: Yazl.ZipFile, zipName: string): Promise<void> {
-        return new Promise((resolve) => {
+    private writeZip(zip: Yazl.ZipFile, zipName: string): Promise<void> {
+        return new Promise((resolve, reject) => {
             fs.mkdirpSync(path.dirname(zipName));
-            zip.outputStream.pipe(fs.createWriteStream(zipName)).on('close', resolve);
+            zip.outputStream.pipe(fs.createWriteStream(zipName))
+                .on('close', resolve)
+                .on('error', reject);
         });
     }
 }
